requirements.txt
streamlit
tensorflow
numpy
tensorflow_text_nightly
tf-nightly